requirements.txt
Django 2.1.3
django-crispy-forms
Pillow

